Add files via upload
As we learn about ANNs and how they can be used in real business processes, we need to become more familiar with common challenges in deploying machine learning models. The following article provides an excellent survey of these challenges by studying a variety of real cases:

What are the main benefits of using ANNs in your organization?
How do ANNs imitate the learning process in the human brain to learn patterns in data?
Can ANNs be used for both classification and regression tasks?
What are the main components of a typical ANN?
Come prepared to discuss the advantages of deep learning over other machine learning models.
Come prepared to discuss in which cases we don’t need to use deep learning.
Ready to answer questions in the article Challenges in Deploying Machine Learning: a Survey of Case Studies.